Fix restart for numberings in different elements.
When adding a new document to a composer the first numbering is restarted. Up to
now we assumed that all numbering parts are added at once, within the same
element. This may not be the case though as there sometimes are intermittent
numberings with e.g. empty paragraphs in between.