Fix handling of section type
The type of a section break is determined by the next section.
The last section always stays at the end. Therefore we need to adjust the
section type of the first new inserted section.
We also need to change the type of the last section of the composed document
to the one from the appended document.

Fixes #10