Fix list-styles being when restarting numberings.
If there is a numId referenced from the paragraph, that numId seems relevant,
otherwise fall back to the style's numId.

The fixture file `switched_listing_style.docx` is replaced as the test asserts
that styles should not be switched, but the output file contained another list
style than the input file. The other fixture files have different xml but when
opening and comparing them in word visually the list styles and numberings
seem to be correct.