Fix docprop update artifacts of cached values.
In addition to always updating the first text node of a complex doc property
with the whole cached value, also remove any additional text-nodes of the
complex doc property. These text-nodes can contain parts of the cached value,
for example when the value is cut into pieces by intermittent proofErr anchors.

This seems to be the correct way to update a complex field,  according to
ECMA-376 5th edition Part 1: §17.16:

  For a complex field represented by use of properly matched fldChar elements
  (§17.16.18), this means replacing any content between the fldChar element
  with an fldCharType attribute value of separate and an fldCharType attribute
  value of end.