Improve update of docproperties.
There were several cases that we did not handle correctly when
updating docproperties, notably:
- Several docproperties in the same paragraph. There was a bug here
leading for example to deletion of text found in between the fields.
- Same docproperty used several times in a document
- Docproperty fields with unquoted property name. This was not handled
at all, leading to the fields not being updated
- Docproperty fields with two spaces before property name. This was not handled
at all, leading to the fields not being updated

We refactor the code in a more object oriented way with classes
representing the docproperties fields. We also improve parsing
using regexes in certain instances instead of string matching,
allowing us to correctly handle different numbers of spaces and
missing quotes.