Update docproperties in header and footer parts.
With the bump of python-docx to version 0.8.8 we no have support
to modify/read headers and footers of word documents. We now also
update doc-properties in headers and footers.

As word seems to keep references to "invisible" header and footer
parts (i.e. when they were previously different for odd/even page,
but that setting has been disabled via word UI) we bypass the official
api provided by docx and check if there is a part referenced in xml.
We always update all parts when they exists as they may "reappear" when
settings in word are changed again.