Fix an issue with utf-8 binary_type docproperties.
We allow any string_types docproperties, this is equal to basestring on
python2 and can include `str` and `unicode`. We don't handle the case
when they are encoded in utf-8 though. This lead to an error when
attempting to set doc-properties that are encoded as utf-8.

I would propose to add separate conditionals for binary and text type,
assuming utf-8 when we receive a binary value and attempting to decode
from utf-8.

This will also still work with ascii docproperties.