Use master header and footer when composing docs with sections.
Because the secPr of the master template is usually in the body
it only applies to the last section of the composed document. This
is fine as long as we only add documents themselves having a single
section, but breaks down as soon as we have multiple sections. We
then need to make sure to add the header and footer to the first
section of the document, so that it gets inherited to the other
sections of the document.