vt2value(): Convert empty <vt:lpwstr/> nodes to empty string instead of None:
This fixes an asymmetry in how values get converted from Python data types to
XML VTs, and vice versa. While empty Python strings are turned into empty
<vt:lpwstr/> nodes, the same was not true the other way round: When reading
docproperties from a document, an empty <vt:lpwstr/> node was turned into
Python `None`.

This asymmetry caused a bug when updating cached docproperty values: In cases
where a docproperty's value in the document's docProps/custom.xml was empty,
it was parsed to Python `None` and therefore the update of the cached value
was skipped in update_all().

That lead to placeholder values in the document never getting replaced with
empty string for those doc properties.