Respect document language when updating docproperties.
Dates can be language dependent as format can be set to display
month and day names.
Determining the correct language to use for a given field would be
difficult, so instead we simply use the first language tag found
in the document or in the styles as language for the whole document.